settings: Remove settings file: begins
